+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- [#43](https://github.com/ForNeVeR/VerifyEncoding/issues/43): new parameter `-ExcludePatterns` accepting an array of glob patterns to exclude files from checks.
+
 ## [2.2.1] - 2025-07-26
 ### Fixed
 - [#29](https://github.com/ForNeVeR/VerifyEncoding/issues/29): Error if a file contains only BOM character. Thanks to @kolosovpetro!